Rename files to be easier to search for
I *heavily* rely on FZF in my development environment, and I've been
frustrated by the collisions between achievement and achievements. So
I'm renaming achievements -> rules, and leaving achievement as the
Achievement/Rule interface.

    rules/
        h001_fixup.rs
        h002_shortest_subject_line.rs
        ...
    achievement/
        achievement.rs
        process_rules.rs
        rule.rs
        ...
