+++ src/lib.rs
@@ -1,8 +1,8 @@
 pub mod achievement;
-pub mod achievements;
 pub mod cli;
 pub mod commands;
 pub mod config;
 pub mod git;
+pub mod rules;
 #[cfg(test)]
 pub mod test;
