+++ src/rules/h002_shortest_subject_line.rs
